RN-1049: Broke ExportDashboard into two separate screens
